+++ starky2/src/lib.rs
@@ -8,6 +8,7 @@ pub mod constraint_consumer;
 pub mod cross_table_lookups;
 mod get_challenges;
 // pub mod mock_stark;
+pub mod all_starks;
 pub mod permutation;
 pub mod proof;
 pub mod prover;
